Fix some record classes and add two new record classes.
git-svn-id: http://npoi.googlecode.com/svn/trunk@281 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ main/HSSF/UserModel/HSSFCellStyle.cs
@@ -497,7 +497,7 @@ namespace NPOI.HSSF.UserModel
             {
                 // Then we need to clone the format string,
                 //  and update the format record for this
-                short fmt = workbook.CreateFormat(
+                short fmt = (short)workbook.CreateFormat(
                         source.GetDataFormatString()
                 );
                 this.DataFormat=(fmt);
